+++ CHANGELOG.md
@@ -4,6 +4,18 @@ Wszystkie istotne zmiany w projekcie `doql`. Format oparty na [Keep a Changelog]
 
 ## [Unreleased]
 
+## [1.0.37] - 2026-06-08
+
+### Docs
+- Update README.md
+
+### Test
+- Update tests/test_parser_benchmark.py
+
+### Other
+- Update app.doql.less
+- Update project/planfile-tickets.yaml
+
 ## [1.0.36] - 2026-06-08
 
 ### Docs